requirements.txt
Django==2.0.3
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
djangorestframework==3.8.2
Pillow==5.1.0
PyJWT==1.6.1
PyMySQL==0.8.0
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
raml-codec==0.1.1
django-rest-raml
coreapi
django-rest-swagger
